backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
sqlalchemy
python-multipart
vosk
pydub
easyocr
pillow
piexif
requests
PyPDF2
python-docx